api/requirements.txt
anyascii==0.2.0
asgiref==3.3.4
astroid==2.5.6
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
dj-stripe==2.5.0
Django==3.2.4
django-allauth==0.44.0
django-appconf==1.0.4
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-imagekit==4.0.2
django-import-export==2.5.0
django-loginas==0.3.9
django-modelcluster==5.1
django-recaptcha==2.0.6
django-redis==5.0.0
django-storages==1.11.1
django-taggit==1.4.0
django-treebeard==4.5.1
djangorestframework==3.12.4
docutils==0.17.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==2.10
isort==5.9.1
jsonfield==3.1.0
l18n==2020.6.1
lazy-object-proxy==1.6.0
mailchimp-marketing==3.0.21
MarkupPy==1.14
mccabe==0.6.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
pilkit==2.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
pylint==2.8.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
stripe==2.58.0
tablib==3.0.0
telepath==0.2
toml==0.10.2
urllib3==1.26.5
wagtail==2.13.2
webencodings==0.5.1
Willow==1.4
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0